BackEnd/pyproject.toml
[project]
name = "snapbooks"
version = "0.1.0"
description = "Telegram AI Accountant for Indian SMBs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "google-genai>=1.0.0",
    "fpdf2>=2.8.0",
    "structlog>=24.0.0",
]
